refactor(*): update package name and license info
