+++ blackbox-no-op/src/main/AndroidManifest.xml
@@ -1,4 +1,4 @@
-<manifest package="com.hr.blackbox">
+<manifest package="dev.anvith.blackbox">
 
 
 </manifest>
