+++ blackbox-no-op/src/main/java/dev/anvith/blackbox/Blackbox.kt
@@ -1,0 +1,22 @@
+package dev.anvith.blackbox
+
+import android.content.Context
+
+/**
+ * Created by Anvith on 09/09/20.
+ */
+class Blackbox {
+
+    companion object {
+        fun context(context: Context): Companion {
+            return this
+        }
+
+        fun otherUncaughtExceptionHandlers(handlers: List<Thread.UncaughtExceptionHandler>): Companion {
+            return this
+        }
+
+        fun init() {
+        }
+    }
+}
