--- blackbox/src/main/java/com/hr/blackbox/data/Constants.kt
@@ -1,14 +1,0 @@
-package com.hr.blackbox.data
-
-/**
- * Created by Anvith on 09/09/20.
- */
-
-object Constants {
-
-    const val NOTIFICATION_ID = 1
-    const val PENDING_REQUEST_ID1 = 101
-    const val PENDING_REQUEST_ID2 = 102
-    const val PENDING_REQUEST_ID3 = 103
-    const val NOTIFICATION_CHANNEL_ID = "blackbox_crash_notification"
-}
