--- blackbox/src/main/java/com/hr/blackbox/data/ExceptionInfo.kt
@@ -1,11 +1,0 @@
-package com.hr.blackbox.data
-
-/**
- * Created by Anvith on 10/09/20.
- */
-
-data class ExceptionInfo(
-    val occurredAt: Long,
-    val name: String,
-    val trace: String
-)
