--- blackbox/src/main/java/com/hr/blackbox/data/StorageProvider.kt
@@ -1,29 +1,0 @@
-package com.hr.blackbox.data
-
-import android.content.Context
-
-/**
- * Created by Anvith on 09/09/20.
- */
-
-interface Storage {
-
-    fun store(exception: ExceptionInfo)
-
-    fun getExceptions(): List<ExceptionInfo>
-
-    fun getLastException(): ExceptionInfo?
-}
-
-class StorageProvider {
-
-    companion object {
-        private lateinit var storage: FileStorage
-        fun getStorage(context: Context): Storage {
-            if (!::storage.isInitialized) {
-                storage = FileStorage(context)
-            }
-            return storage
-        }
-    }
-}
