+++ blackbox/src/main/java/dev/anvith/blackbox/data/Constants.kt
@@ -1,0 +1,14 @@
+package dev.anvith.blackbox.data
+
+/**
+ * Created by Anvith on 09/09/20.
+ */
+
+object Constants {
+
+    const val NOTIFICATION_ID = 1
+    const val PENDING_REQUEST_ID1 = 101
+    const val PENDING_REQUEST_ID2 = 102
+    const val PENDING_REQUEST_ID3 = 103
+    const val NOTIFICATION_CHANNEL_ID = "blackbox_crash_notification"
+}
