+++ sample/src/main/AndroidManifest.xml
@@ -1,16 +1,18 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.hr.blackbox">
+    xmlns:tools="http://schemas.android.com/tools"
+    package="dev.anvith.blackbox">
 
     <application
+        android:name=".BlackboxApplication"
         android:allowBackup="true"
         android:icon="@mipmap/ic_launcher"
         android:label="@string/app_name"
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
-        android:name=".BlackboxApplication"
-        android:theme="@style/AppTheme">
-        <activity android:name=".MainActivity">
+        android:theme="@style/AppTheme"
+        tools:ignore="AllowBackup">
+        <activity android:name="dev.anvith.blackbox.MainActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
