--- sample/src/main/java/com/hr/blackbox/BlackboxApplication.kt
@@ -1,17 +1,0 @@
-package com.hr.blackbox
-
-import android.app.Application
-
-/**
- * Created by Anvith on 08/09/20.
- */
-class BlackboxApplication : Application() {
-    override fun onCreate() {
-        super.onCreate()
-        if (BuildConfig.DEBUG) {
-            //initialise reporter with external path
-            Blackbox.context(this)
-                .init()
-        }
-    }
-}
